Known Patch from NCAR 2016-11-30
Problem with Cloud Fraction (posted November 30, 2016)

Problem: Beginning in V3.8, a bug was introduced to cloud fraction.
This is seen by most radiation schemes, when a microphysics scheme is
used that does not have cloud ice (e.g. WSM3, Kessler). When WSM3 or
Kessler is used, cloud fraction is not computed, and hence a radiation
scheme will not see any cloud. This results in unrealistic surface
temperature and incorrect upwelling long-wave radiation at the top of
the atmosphere.

Solution: To correct the problem, download this modified
module_radiation_driver.F.fix-for-v3.8.1.tar.gz file. You will need to
unpack the file and place it in the WRFV3/phys/ directory. Rename the
file to 'module_radiation_driver.F' and recompile the code. It will not
be necessary to issue a 'clean -a' or to reconfigure. Simply recompile,
and the compilation should be rather quick.